requirements.txt
# ============================================================================
# Web framework
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================================
# Database
# ============================================================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# ============================================================================
# Task Queue
# ============================================================================
celery==5.3.4
redis==5.0.1

# ============================================================================
# Validation & Settings
# ============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator

# ============================================================================
# Auth
# ============================================================================
PyJWT==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# LLM / AI
# ============================================================================
ollama
httpx
requests
tenacity

# ============================================================================
# Media: Image / Video / Audio
# ============================================================================
Pillow
ffmpeg-python
pydub

# ============================================================================
# Web scraping & social APIs
# ============================================================================
beautifulsoup4
feedparser
tweepy
praw

# ============================================================================
# Logging & monitoring
# ============================================================================
structlog
python-json-logger
prometheus-client

# ============================================================================
# Utilities
# ============================================================================
python-slugify
pytz
tzlocal
alembic
feedparser
ffmpeg-python
httpx
ollama
passlib[bcrypt]
Pillow
praw
pydub
python-jose[cryptography]
python-json-logger
python-slugify
structlog
tenacity
tweepy
tzlocal
